# docs/swagger.yaml
basePath: /v1
definitions:
  entity.Appointment:
    properties:
      appointment_time:
        type: string
      created_at:
        type: string
      doctor_id:
        type: integer
      duration:
        description: in minutes
        type: integer
      id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Appointment:
    type: object
  models.AppointmentResponse:
    properties:
      doctor_id:
        type: integer
      id:
        type: integer
      user_id:
        type: integer
    type: object
  models.AppointmentsResponse:
    properties:
      appointments:
        items:
          $ref: '#/definitions/entity.Appointment'
        type: array
    type: object
...
  models.Error:
    properties:
      message:
        type: string
    type: object
...
  models.SignInUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  models.SignUpUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - password
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      phone:
        type: string
    type: object
host: localhost:8070
info:
  contact: {}
  description: API for doctor appointment
  title: Doctor appointment api
  version: "1.0"
paths:
  /appointments:
    post:
      consumes:
      - application/json
      description: Create appointment
      parameters:
      - description: Appointment
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Create appointment
      tags:
      - appointment
    put:
      consumes:
      - application/json
      description: Update appointment
      parameters:
      - description: Appointment
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/models.Appointment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Update appointment
      tags:
      - appointment
  /appointments/{appointment_id}:
    delete:
      consumes:
      - application/json
      description: Delete appointment
      parameters:
      - description: Appointment ID
        in: path
        name: appointment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Delete appointment
      tags:
      - appointment
    get:
      consumes:
      - application/json
      description: Get appointment by id
      parameters:
      - description: Appointment ID
        in: path
        name: appointment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get appointment by id
      tags:
      - appointment
  /appointments/doctor/{doctor_id}:
    get:
      consumes:
      - application/json
      description: Get appointments by doctor id
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get appointments by doctor id
      tags:
      - appointment
  /appointments/doctor/{doctor_id}/booked-schedules:
    get:
      consumes:
      - application/json
      description: Get booked schedules by doctor id
      parameters:
      - description: Doctor ID
        in: path
        name: doctor_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppointmentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get booked schedules by doctor id
      tags:
      - appointment
  /appointments/user/{user_id}:
    get:
      consumes:
      - application/json
      description: Get appointments by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppointmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get appointments by user id
      tags:
      - appointment
  /appointments/user/{user_id}/booked-schedules:
    get:
      consumes:
      - application/json
      description: Get booked schedules by user id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AppointmentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Get booked schedules by user id
      tags:
      - appointment
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Sign in a user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SignInUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Sign in user
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Sign up a user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.SignUpUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
      summary: Sign up user
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"